launch/pyproject.toml
[project]
name = "launch"
version = "0.1.0"
description = "RepoLaunch"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "docker>=7.1.0",
    "fire>=0.7.0",
    "gitpython>=3.1.44",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "tenacity>=9.1.2",
    "tornado>=6.5.1",
    "typing-extensions>=4.14.1",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["launch"]
